Handle text_input_cursor_rect_for_byte_offset
