+++ internal/renderers/femtovg/fonts.rs
@@ -3,7 +3,7 @@
 
 // cspell:ignore Noto fontconfig
 
-use crate::PhysicalLength;
+use crate::{PhysicalLength, PhysicalPoint};
 use core::num::NonZeroUsize;
 use femtovg::TextContext;
 use i_slint_common::sharedfontique::{self, parley};
@@ -64,4 +64,35 @@ pub fn layout(
         parley::AlignmentOptions::default(),
     );
     layout
+}
+
+pub fn get_cursor_location(
+    layout: &parley::Layout<()>,
+    cursor_byte_offset: usize,
+    offset: f32,
+) -> Option<PhysicalPoint> {
+    let mut cursor_point = None;
+
+    for line in layout.lines() {
+        for item in line.items() {
+            match item {
+                parley::PositionedLayoutItem::GlyphRun(glyph_run) => {
+                    let range = glyph_run.run().text_range();
+                    if range.contains(&cursor_byte_offset) {
+                        cursor_point = glyph_run
+                            .positioned_glyphs()
+                            .nth(cursor_byte_offset - range.start)
+                            .map(|glyph| PhysicalPoint::new(glyph.x, glyph.y + offset));
+                    } else if cursor_byte_offset == range.end {
+                        cursor_point = glyph_run.positioned_glyphs().last().map(|glyph| {
+                            PhysicalPoint::new(glyph.x + glyph.advance, glyph.y + offset)
+                        });
+                    }
+                }
+                parley::PositionedLayoutItem::InlineBox(_inline_box) => {}
+            };
+        }
+    }
+
+    cursor_point
 }
